+++ rpc/grpc_test.go
@@ -1,8 +1,10 @@
 package rpc
 
 import (
+	"fmt"
 	"log/slog"
 	"net"
+	"sync/atomic"
 	"testing"
 	"time"
 
@@ -19,24 +21,8 @@ import (
 )
 
 func TestGRPCIntegration(t *testing.T) {
-	server := grpc.NewServer(
-		grpc.MaxConcurrentStreams(1),
-	)
-	service := mocks.RPCServer{}
-	pb.RegisterRPCServer(server, &service)
-
-	listen, err := net.Listen("tcp", ":50051") // nolint: gosec
-	if err != nil {
-		t.Fatalf("failed to listen: %v", err)
-	}
-
-	defer server.GracefulStop()
-
-	go func() {
-		if err := server.Serve(listen); err != nil {
-			t.Errorf("failed to serve: %v", err)
-		}
-	}()
+	service, cleanup := runGRPCServer(t, 50051)
+	defer cleanup()
 
 	// Connect is fast and successful
 	service.On("Connect", mock.Anything, mock.Anything).Return(&pb.ConnectionResponse{Status: pb.Status_SUCCESS}, nil)
@@ -114,4 +100,66 @@ func TestGRPCIntegration(t *testing.T) {
 		assert.Equal(t, rpcCalls+1, metrics.Counter(metricsRPCCalls).Value())
 		assert.Equal(t, rpcRetries+1, metrics.Counter(metricsRPCRetries).Value())
 	})
+}
+
+func TestGRPCMultiHost(t *testing.T) {
+	service, cleanup := runGRPCServer(t, 50151)
+	defer cleanup()
+
+	service2, cleanup2 := runGRPCServer(t, 50152)
+	defer cleanup2()
+
+	var service1_called int32
+	var service2_called int32
+
+	service.On("Connect", mock.Anything, mock.Anything).Return(&pb.ConnectionResponse{Status: pb.Status_SUCCESS}, nil).Run(func(args mock.Arguments) {
+		atomic.AddInt32(&service1_called, 1)
+	}).After(10 * time.Microsecond)
+
+	service2.On("Connect", mock.Anything, mock.Anything).Return(&pb.ConnectionResponse{Status: pb.Status_SUCCESS}, nil).Run(func(args mock.Arguments) {
+		atomic.AddInt32(&service2_called, 1)
+	}).After(10 * time.Microsecond)
+
+	m := metrics.NewMetrics(nil, 0, slog.Default())
+	l := slog.Default()
+	headers := map[string]string{"cookie": "token=secret;"}
+
+	c := NewConfig()
+	c.Host = "grpc-list://localhost:50151,localhost:50152"
+	controller, err := NewController(m, &c, l)
+	require.NoError(t, err)
+	require.NoError(t, controller.Start())
+	defer controller.Shutdown() //nolint:errcheck
+
+	for i := 0; i < 10; i++ {
+		res, err := controller.Authenticate("test", &common.SessionEnv{URL: "http://test.cable", Headers: &headers})
+
+		require.NoError(t, err)
+		require.NotNil(t, res)
+		assert.Equal(t, common.SUCCESS, res.Status)
+	}
+
+	assert.Greater(t, service1_called, int32(0))
+	assert.Greater(t, service2_called, int32(0))
+}
+
+func runGRPCServer(t *testing.T, port int) (*mocks.RPCServer, func()) {
+	server := grpc.NewServer(
+		grpc.MaxConcurrentStreams(1),
+	)
+	service := mocks.RPCServer{}
+	pb.RegisterRPCServer(server, &service)
+
+	listen, err := net.Listen("tcp", fmt.Sprintf(":%d", port)) // nolint: gosec
+	if err != nil {
+		t.Fatalf("failed to listen: %v", err)
+	}
+
+	go func() {
+		if err := server.Serve(listen); err != nil {
+			t.Errorf("failed to serve: %v", err)
+		}
+	}()
+
+	return &service, func() { server.GracefulStop() }
 }